requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
groq==0.13.1
pydantic==2.10.4
python-dotenv==1.0.1
streamlit==1.41.1
